Fix segmentation fault when handling packets larger than MTU + Timeout minor fix
This commit fixes the switch segmenation fault when packets
larger than the interface MTU are received.

The problem is caused mainly due to TCP segmentation offload, which
enables the TCP stack to send an entire chunk of data to be break up by
the network interface.

While the switch does not break anymore, it still does not handle TSO,
so larger packets might be dropped when sent to the switch causing
loss of performance. For this reason is still recommended to disable
TSO support on the interface.

Conflicts:
	udatapath/dp_ports.c